# Calculadora/swagger.yaml
swagger: "2.0"
info:
  description: "Servicio que permite realizar operaciones matemáticas (suma, resta, multiplicación,división)"
  version: "1.0.0"
  title: "Servicio Calculadora"
host: "localhost:5000"
schemes:
- "http"
paths:
  /sumar/:
    post:
      description: Retorna la suma de los números indicados
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Lista de numeros separada por :"
        required: true
        schema:
          $ref: "/schema/numeros.json"
      responses:
        200:
          description: Retorna el valor de la suma de los números.
          schema:
            $ref: "/schema/resultado.json" 
        404:
          description: Error en los parámetros, página no encontrada.
  /restar/:
    post:
      description: Retorna la resta de dos números indicados
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Números a restar"
        required: true
        schema:
          $ref: "/schema/numerosrd.json"
      responses:
        200:
          description: Retorna el valor de la resta de los números.
          schema:
            $ref: "/schema/resultado.json" 
        404:
          description: Error en los parámetros, página no encontrada.
  /multiplicar/:
    post:
      description: Retorna la multiplicación de los números indicados 
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Lista de numeros separada por :"
        required: true
        schema:
          $ref: "/schema/numeros.json"
      responses:
        200:
          description: Retorna el valor de la multiplicación de los números.
          schema:
            $ref: "/schema/resultado.json" 
        404:
          description: Error en los parámetros, página no encontrada.
  /dividir/:
    post:
      description: Retorna la división de dos números indicados
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Números a dividir"
        required: true
        schema:
          $ref: "/schema/numerosrd.json"
      responses:
        200:
          description: Retorna el valor de la división de los números.
          schema:
            $ref: "/schema/resultado.json" 
        404:
          description: Error en los parámetros, página no encontrada.
        